LC-construct-product-matrix [Runtime: 26 ms; Memory: 131.8 MB]

+++ array/construct-product-matrix.cpp
@@ -8,7 +8,7 @@ public:
 
         // result[i] = (product of left) * (product of right)
 
-        int prefix = 1;
+        long long prefix = 1;
         for(int i=0; i<m; i++){
             for(int j=0; j<n; j++){
                 prodMat[i][j] = prefix;
@@ -16,7 +16,7 @@ public:
             }
         }
 
-        int suffix = 1;
+        long long suffix = 1;
         for(int i=m-1; i>=0; i--){
             for(int j=n-1; j>=0; j--){
                 prodMat[i][j] = (prodMat[i][j] * suffix) % mod;
